refactor(ui): split MODELS into its own panel
Sorotassu preferred this over the sub-section-title treatment from
round 2. Models is not really a sub-section of the run options —
it is a distinct concern (checkpoint management) that happens to live
in the same UI. Splitting matches how the app actually thinks about
these two responsibilities.

- New ModelsPanel.tsx owns the model download state machine, HF
  token input, /model-status polling, and per-model error inline
  text. Renders as its own .container with the standard uppercase
  .headerTitle treatment ('MODELS') matching OPTIONS / PROGRESS.
- OptionsPanel loses the models block: 66 lines out, plus the
  DownloadState type, refreshModelStatuses, downloadModel,
  buttonLabel, buttonClass helpers, and the four related useState
  slots. Also drops the now-unused 'connected' prop.
- Right sidebar layout: options fills flex: 1 min-height: 0
  (needed with the container fixed to overflow: hidden), models
  sizes to content via flex: 0 0 auto, gap: 1rem between them.

Reverts the round-2 .subSectionTitle usage for Models; keeps the CSS
rule available for legitimate future sub-sections.
